Refactor ledger utils
